Add files via upload
Fixed the bug with gameOver places. Now the board is being fixed when created or reseted using fixTheBoard() func. gameOver areas are eliminated by replacing a closed wall of gameOver area with one intersection symbol (ASCII 206). gameOver() func is no longer needed but I'll let it stay in the code.


Next version of fully completed game but with few bugs:

- pacman sometimes leaves poop of himself. No clue why. If solution not found - replace pacMan instance with one of the boards symbols (nobody will notice probably but it is a shitty solution)

- test multiplayer

- there is often a situation when a player gets the berry in the last second and at the same time the berry is eaten but the time runs out insted of add 3 more seconds

- clean the code